v0.0.1
This is the beginning of my workout tracker. Finding workouts online
that i enjoy doing I want to find the best way to store them and then
randomly serve them up when I’m looking for a workout to complete.

In 0.0.1 I’ve set up the first page and added in some css. I will be
building on this app every week while I learn how to code. Right now
the content box is not linked to anything and won’t store on the
content you submit.